more interop testing
